requirements.txt
streamlit==1.32.0
pandas==2.2.1
snowflake-connector-python==3.7.0
boto3==1.34.69
plotly==5.19.0
python-dotenv==1.0.1
numpy==1.26.4
pytest==8.0.2 